requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.3
bitarray==3.7.1
bitstring==4.3.1
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
esptool==5.1.0
et_xmlfile==2.0.0
Faker==20.1.0
forex-python==1.9.2
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
intelhex==2.3.0
itypes==1.2.0
Jinja2==3.1.6
kafka-python==2.0.2
kombu==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.0.7
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
prometheus-client==0.19.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyserial==3.5
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
reedsolo==1.7.0
requests==2.32.5
rich==14.1.0
rich-click==1.9.1
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
